refactor : fixture, constants로 변경
